scv-24: adding criminal setences table
