+++ web/src/components/criminal/CriminalSentenceDetails.vue
@@ -1,21 +1,5 @@
 <template>
-   <b-card  v-if= "isMounted" no-body>
-        <div>         
-            <h3 class="mx-2 mt-5 font-weight-normal"> Counts ({{NumberOfCounts}}) </h3>
-            <hr class="mx-2 bg-light" style="height: 5px;"/>         
-        </div>      
-        
-        <b-card>
-            <b-dropdown  variant="light text-info" :text="getNameOfParticipant(activeCriminalParticipantIndex)" class="m-0">    
-                <b-dropdown-item-button  
-                    v-for="participant in SortedParticipants" 
-                    :key="participant['Index']"
-                    v-on:click="setActiveParticipantIndex(participant['Index'])">
-                        {{getNameOfParticipant(participant['Index'])}}
-                </b-dropdown-item-button> 
-            </b-dropdown> 
-        </b-card>
-        <b-card> 
+   <b-card  v-if= "isMounted" no-body> 
         <b-table-simple small responsive borderless>
             <b-thead>                    
                 <b-tr >
@@ -74,8 +58,7 @@
                     
                 </b-tr>                       
             </b-tbody>
-        </b-table-simple>
-      </b-card>       
+        </b-table-simple>      
    </b-card> 
 </template>
 
@@ -89,22 +72,17 @@ const criminalState = namespace('CriminalFileInformation');
 @Component
 export default class CriminalSentenceDetails extends Vue {
 
-    @criminalState.State
-    public criminalFileInformation!: any
-
-    @criminalState.Action
-    public UpdateCriminalFile!: (newCriminalFileInformation: any) => void
+   
 
     @criminalState.State
-    public activeCriminalParticipantIndex    
+    public criminalParticipantSentenceInformation  
 
-    @criminalState.Action
-    public UpdateActiveCriminalParticipantIndex!: (newActiveCriminalParticipantIndex: any) => void
-
+    
     public getParticipants(): void {       
-        const data = this.criminalFileInformation.detailsData;
-        this.participantJson = data.participant
-        this.ExtractParticipantInfo()          
+         
+        this.participantFiles = this.criminalParticipantSentenceInformation.participantFiles
+        this.selectedParticipant = this.criminalParticipantSentenceInformation.selectedParticipant
+         
         this.isMounted = true;
     }
 
@@ -118,6 +96,8 @@ export default class CriminalSentenceDetails extends Vue {
    
     isMounted = false
 
+    selectedParticipant = 0;
+
     participantFiles: any[] = [];
 
     fields = [        
@@ -130,145 +110,18 @@ export default class CriminalSentenceDetails extends Vue {
         {key:'Amount',                   sortable:false,  tdClass: 'border-top', headerStyle:'text',          cellStyle:'text'},
         {key:'Due Date/ Until',          sortable:false,  tdClass: 'border-top', headerStyle:'text',          cellStyle:'text'},
         {key:'Effective Date',           sortable:false,  tdClass: 'border-top', headerStyle:'text',          cellStyle:'text'},
-    ];
-
-    public getNameOfParticipant(num)
-    {        
-        if(!this.participantFiles[num]["First Name"])
-            return  this.participantFiles[num]["Last Name"];
-        else if(!this.participantFiles[num]["Last Name"])
-            return this.participantFiles[num]["First Name"];
-        else
-            return  this.participantFiles[num]["Last Name"]+', '+this.participantFiles[num]["First Name"];           
-    }
-
-    public setActiveParticipantIndex(index)
-    {                   
-        this.UpdateActiveCriminalParticipantIndex(index);  
-    }	
-    
-    public ExtractParticipantInfo(): void {        
-        
-        for(const fileIndex in this.participantJson)
-        {            
-            const fileInfo = {};
-            const jFile =  this.participantJson[fileIndex];
-            fileInfo["Index"] = fileIndex; 
-            fileInfo["Part ID"] = jFile.partId;
-            fileInfo["First Name"] = jFile.givenNm.trim().length>0 ? jFile.givenNm : "";
-            fileInfo["Last Name"] = jFile.lastNm ? jFile.lastNm : jFile.orgNm;            
-            
-            fileInfo["Counts"] = [];
-            const counts: any[] = [];           
-              
-            for(const cnt of this.mergeSentences(jFile.count))           
-            {                
-                const countInfo = {}; 
-               
-                countInfo["Date"] = cnt.appearanceDate? cnt.appearanceDate.split(' ')[0] : ''; 
-                                
-                countInfo["Finding"]= cnt.finding? cnt.finding:'';
-                countInfo["FindingDsc"]= cnt.findingDsc? cnt.findingDsc:'';
-                                   
-                countInfo["Count"]='';
-                countInfo["ChargeIssueCd"] = [] ;
-                countInfo["ChargeIssueDsc"] = [];
-                countInfo["ChargeIssueDscFull"] = [];
-                
-                for(const charge of cnt.charge)
-                { 
-                    countInfo["Count"] += charge.countNum + ',';
-                    countInfo["ChargeIssueCd"].push(charge.chargeTxt? charge.chargeTxt: '') ;
-                    countInfo["ChargeIssueDsc"].push(charge.chargeDscTxt? (charge.chargeDscTxt.length>10 ? charge.chargeDscTxt.substr(0,10)+' ...':charge.chargeDscTxt): '');
-                    countInfo["ChargeIssueDscFull"].push(charge.chargeDscTxt? charge.chargeDscTxt:'');
-                }
-                countInfo["Count"]= countInfo["Count"].slice(0, -1); 
-
-                countInfo["Sentence/Disposition Type"]=[];
-                countInfo["SentenceDsc"]=[];
-                countInfo["Term"] = [];
-                countInfo["Amount"]=[];
-                countInfo["Due Date/ Until"]=[];
-                countInfo["Effective Date"]=[]; 
-
-                countInfo["LenCharge"] = cnt.charge.length;
-                countInfo["Len"]= Math.max(cnt.sentence.length , countInfo["LenCharge"]); 
-               
-                for(const sentence of cnt.sentence)
-                {                    
-                   countInfo["Sentence/Disposition Type"].push(sentence.sntpCd); 
-                   countInfo["SentenceDsc"].push(sentence.sentenceTypeDesc? sentence.sentenceTypeDesc:'');
-                   countInfo["Term"].push(sentence.sentTermPeriodQty? (sentence.sentTermPeriodQty + ' ' + sentence.sentTermCd.replace('-','')):'')
-                   countInfo["Amount"].push(sentence.sentMonetaryAmt? sentence.sentMonetaryAmt:'')
-                   countInfo["Due Date/ Until"].push(sentence.sentDueTtpDt? sentence.sentDueTtpDt.split(' ')[0]:'')
-                   countInfo["Effective Date"].push(sentence.sentEffectiveDt?  sentence.sentEffectiveDt.split(' ')[0]:'')                   
-                }                
-                
-                if(cnt.sentence.length < countInfo["Len"])
-                {
-                    for(let loop=0; loop < (countInfo["Len"]-cnt.sentence.length)  ;loop++)
-                    {                    
-                        countInfo["Sentence/Disposition Type"].push(''); 
-                        countInfo["SentenceDsc"].push('');
-                        countInfo["Term"].push('')
-                        countInfo["Amount"].push('')
-                        countInfo["Due Date/ Until"].push('')
-                        countInfo["Effective Date"].push('')                   
-                    } 
-                }
-                else if(countInfo["LenCharge"] < countInfo["Len"])
-                {
-                    for(let loop=0; loop < (countInfo["Len"]-countInfo["LenCharge"])  ;loop++)
-                    {       
-                        countInfo["ChargeIssueCd"].push('');
-                        countInfo["ChargeIssueDsc"].push('');
-                        countInfo["ChargeIssueDscFull"].push('');
-                    } 
-                }
-                  
-                counts.push(countInfo);
-            }
-            fileInfo["Counts"] = counts;
-                        
-            this.participantFiles.push(fileInfo);
-        } 
-    
-    }
-
-    public mergeSentences(counts) 
-    {
-        const groupedCounts = _.groupBy(counts, function (count: any) {
-            const orderedSentences = _.sortBy(count.sentence, "sntpCd");                
-            return  count.appearanceDate +count.finding + _.pluck(orderedSentences, "sntpCd") + _.pluck(orderedSentences, "sentTermPeriodQty") + _.pluck(orderedSentences, "sentTermCd")
-                + _.pluck(orderedSentences, "sentMonetaryAmt") + _.pluck(orderedSentences, "sentDueTtpDt") + _.pluck(orderedSentences, "sentEffectiveDt");
-        });
-
-        return _.map(groupedCounts, function (countGroup) {
-            const mergedCount = countGroup[0];
-            mergedCount.charge = _.chain(countGroup)
-            .sortBy(function(stooge){ return stooge.countNumber; })
-            .map(function (group) { return group.countNumber+ "|" + group.sectionTxt + "|" + group.sectionDscTxt })
-            .uniq()
-            .map(function (charge) { return { countNum:charge.split("|")[0], chargeTxt: charge.split("|")[1], chargeDscTxt: charge.split("|")[2] }; })
-            .value();                return mergedCount;
-        });
-    }
+    ];   
     
     get SortedParticipantFilesCounts()
     {
         if(this.dateSortDir =='desc')
-            return _.sortBy(this.participantFiles[this.activeCriminalParticipantIndex]['Counts'],this.sortBy).reverse()
+            return _.sortBy(this.participantFiles[this.selectedParticipant]['Counts'],this.sortBy).reverse()
         else
-            return _.sortBy(this.participantFiles[this.activeCriminalParticipantIndex]['Counts'],this.sortBy)  
+            return _.sortBy(this.participantFiles[this.selectedParticipant]['Counts'],this.sortBy)  
     } 
 
     get NumberOfCounts() { 
-        return(this.participantFiles[this.activeCriminalParticipantIndex]["Counts"].length) 
-    }
-
-    get SortedParticipants()
-    {         
-        return _.sortBy(this.participantFiles,(participant=>{return (participant["Last Name"]? participant["Last Name"].toUpperCase() : '')}))       
+        return(this.participantFiles[this.selectedParticipant]["Counts"].length) 
     }
 
     public getRowStyle(index) {
